% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNestBin.R
\name{rNestBin}
\alias{rNestBin}
\title{Random nested exchangeable binary numbers}
\usage{
rNestBin(means, rhoC, rhoCT, n = 1, C = 1, sample = TRUE)
}
\arguments{
\item{means}{Vector of prevalences, one for each subcluster within the cluster. The length of this vector determines the number of subclusters per cluster.}

\item{rhoC}{Correlation between two observations in the same cluster, but not the same subcluster.}

\item{rhoCT}{Correlation between two observations in the same subcluster.}

\item{n}{Number of observations per subcluster.}

\item{C}{Number of clusters to sample.}

\item{sample}{If sample=FALSE then just check whether the values of rhoC, rhoCT, means are feasible and stop if they are not feasible.}
}
\description{
This function allows you to sample multivariate binary random variables with
a nested exchangeable correlation matrix, and different prevalences (means)
for each sub-cluster.
}
\details{
The output data are binary random variables separated into clusters. Clusters
are independent, and each cluster has the same multivariate joint
distribution as all others. Each cluster is separated into C subclusters.
There are two levels of correlation: between subcluster and within
subcluster. All observations in the same subcluster have the same
prevalence, but different subclusters can have different prevalences.

The elements of the correlation matrix are given by
R\link{i,j} = 1 if i = j
= rhoCT if i and j are observations in the same subcluster
= rhoC if i and j are observations in the same cluster but not the
same subcluster. rhoC <= rhoCT
}
\examples{
rNestBin(rhoC=0.2,rhoCT=0.4,means=c(0.4,0.1),n=20,C=10,sample=T)
}
